creates own class for getting the ipa of sentences
